+++ tools/flexbot-cli/flexbot.go
@@ -19,7 +19,7 @@ import (
 )
 
 const (
-	version = "1.8.1"
+	version = "1.8.2"
 )
 
 // OperationResult interface