+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## 1.8.2 (November 18, 2022)
+## 1.8.2 (November 29, 2022)
 
 ENHANCEMENTS:
-* Multiple improvements in ONTAP REST API client based on OnTap v9.12
-  * Moved LunCopy routine from PrivateCli /storage/luns REST API
+* Multiple improvements in ONTAP REST API client
+  * api_method=`rest` is stable now, requires OnTap v9.12.1 or higher
 
 ## 1.8.1 (October 18, 2022)